PCB made, might check, then gerber if not..
Here is the PCB for the matrix pcb with 8x8 leds, the input pins and output pins looks the same but one will be with socket.  Both female and male connector pins are the same according to the seller / elfa. 
Pin size for male connector = Pin size for femlae connector. 

And the cupper traces could probably been better made x)