Order List Done
Added all the things to the order list see text link and column Mini Audio Spectrum for all the components. If wanted add more to the list
just write them here then again in the google sheet. Don't forget to update the total amount of things / not specific amount just the types so it's easier to check the google if something is missing :).